Add a condinit selection routine (#67)
<!-- Thank you so much for your contribution! -->

<!-- First describe what your PR is doing. -->

This PR propose a new "patchless" way of dealing with condinit.
Everyone can write their own condinit routine on the primitive variable
and add them in the condinit selector.
This avoids teaching people the bad habit of having patches and limit
code duplication.

The tests using patches were converted to this new way.


> [!note]
>  This PR has no breaking changes (the old patch way will still work)

<!-- Please check the following points -->
## PR Checklist

- [x] Pull request points to the `dev` branch
- [x] The pull request has a description
- [x] If it is a bugfix that should be urgently merged to stable, the PR
should have label `on-merge: backport-to-stable`.
- [x] For new code features, the PR should have been tested (please
include the results of the tests below or, when possible, add a new test
for the feature).
- [x] The code covered by the pull request is documented

---------

Co-authored-by: Noe Brucy <noe.brucy@cea.fr>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>