Add coverage statistics to test suite using gcov (#61)
Added an option -s (for statistics) to run the test suite with gcov.
Statistics will be gathered for each test separately and temporarily
stored in the respective test directory. At the end, a script will
combine the statistics from all tests into one folder called coverage.
This folder contains a file for each source file that was compiled,
taking care of duplicate file names. A report is printed to file with
the percentage of lines that was executed for each file, as well as the
overall coverage (for source files compiled by the test suite, i.e. this
currently excludes aton/).

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Corentin Cadiou <contact@cphyc.me>
Co-authored-by: Noé Brucy <nbrucy@users.noreply.github.com>