Add test to cover passive scalar code (#86)
![Screenshot from 2025-03-14
14-50-55](https://github.com/user-attachments/assets/0cba1006-8920-4c15-b272-6d378ab49440)

Description of the test:
A diamond shape is placed in a lower density medium which is moving to
the left. The passive scalar is set to be 1 in the diamond shape and -1
in the background medium. Turbulence is generated in the wake at the
corners of the diamond,
causing the passive scalar to mix.

Reference solution generated using 2 mpi processes.
Execution time on 4 cpus is about 40 seconds on my machine.
I needed to increase the tolerance because the chaotic nature of the
problem makes the solution sensitive to the number of processes used.