Fix underflows in ramses-rt - comp_table_rates function (#68)
The function `comp_table_rates` produces a lot of underflows by design,
with computation such as
`exp(-157809.1d0/T)` with T starting from 0.01.

This is probably not a problem in everyday's life, as these get ignored
as nan or 0, but it start being annoying when compiling in DEBUG mode to
solve real problems elsewhere in the code.

With this PR, I try to detect underflows before they happens and put 0
instead of nan.



[!WARNING]
> The cooling tables are very slightly changed:
> The max difference is 4.435031475934927e-30 and max relative difference (with 0 values taken out) is 9.387004739470413e-12