[tests] Add a -r option to test restart (#110)
Adds an option to interject a restart in all tests. The option is NOT
activated by default on the CI.


Some tests are failing because RAMSES is not binary exact. A solution
would be to store conservative variable instead of primitive ones (as
done in RAMSES ism) (see #123).
There is a more serious issue with the MC tracer test (see #121)

---------

Co-authored-by: Noe Brucy <noe.brucy@ens-lyon.fr>