[tests] fix small values being rounded to 0 by visu_ramses (#127)
Changed the part in visu_ramses which sets small numbers to 0. This was
causing unit_d to be set to 0 in the report for some tests. This issue
and solution were discovered by Matthias Gonzalez and Noe Brucy.
Additionally, this caused RT scalar variables to be artificially set to
0 in reference solutions, when they are small.
Even though in principle they should be 0, we do want to take into
account these errors when testing.
We only keep the fix for sink vector variables: l, v and v gas, since
these errors are non-deterministic and may originate from the csv
reading mechanism.

This PR updates the reference solution of the affected tests.

co-authored by: @matthiasglz , @nbrucy

---------

Co-authored-by: Matthias González <96419241+matthiasglz@users.noreply.github.com>
Co-authored-by: Noé Brucy <noe.brucy@ens-lyon.fr>