[SPACE] Optimize and document nbor utils (#108)
**Refactor routines for obtaining the 3cube neighbors**
* The routine get3cubefather is split into its two parts. Before,
get3cubefather returned both the _cells_ and the _grids_, but only one
of them was actually used after the call. Now, get3cubefather only
returns the neighbor _cells_. A new subroutine get3cubefather_grids,
returns the grids (only needed for poisson multigrid). This avoid
unnecessary computations and memory access.
* Move the first part of the routine get3cubepos to its own routine,
named get_grids_of_nbor_cells, so it can be used by
get3cubefather_grids.
* The input ind for get3cubepos and get_grids_of_nbor_cells is now an
array of length nvector instead of an integer. This allows to remove the
if condition inside the loop over cells to select the cells that sit at
position ind, which hinders vectorization and creates more loop
iterations:
```
        iok=0
        do i=1,ncell
           if(pos(i)==ind)then
```
* Use merge instead of if-else in get3cubepos, in an attempt to improve
vectorization.

These changes reduce the cost of get3cubefather itself to almost
nothing. The computational cost is now purely in the subroutines called
by get3cubefather, rather than the loops at the end of the routine which
were inefficient.
I've measured **performance gains** for several benchmarks:
* cosmo on uniform grid: marenostrum intel 6.2%, meluxina gnu 10.5%
* cosmo with amr:  marenostrum 8.2%, meluxina 11.2%
* sedov on uniform grid: 1-2%

**Documentation**
* Add an overview of the routines at the top of the file.
* Add more detailed descriptions of what each routine is doing.
* Try to make variable naming more intuitive/consistent.

This development was done in the context of the SPACE CoE.